Cliente_Back/requirements.txt
flask==3.0.3
pycryptodome==3.20.0
flask-cors==3.0.10
flask-socketio==5.3.2
requests==2.31.0